Main

Description

Module contains two versions of communication with user. The version should be estimated before running the project. The first option (by implication) is mainWithMessages, which enables communication and instrucing user. The second option is mainWithoutMessages, with this option you can work without neccessary coments. This option might be helpful when you are going to use piping in console.

Some functions may use type Val = Int to differentiate values stored in SegmentTree from Int used to specify boundaries and size.

Documentation

main :: IO ()

This is main function. In this function the appropriate version of cummication with user is called.

mainWithMessages :: IO ()

This function communicates with user and instructs them.

mainWithoutMessages :: IO ()

This function communicates with user without instruction.

interactWithUser :: Int -> SegmentTree -> IO ()

This function proceeds users' commands.