SegmentTree

Description

Module delivers complete structure of semgent tree with listed below useful operations:

set b e v - change values on interval [b,e) into v

add b e v - add value v to all numbers on interval [b,e)

sum b e - return sum of numbers on interval [b,e)

min b e - return minimum number on interval [b,e)

max b e - return maximum number on interval [b,e)

get b e - print numbers on interval [b,e) in list format

Some functions may use type Val = Int to differentiate values stored in SegmentTree from Int used to specify boundaries and size.

Documentation

data SegmentTree

SegmentTree is either NullLeaf or Node containing two children, own boundaries and values to compute answers for queries effectively

Constructors

Null
NullLeaf

Node SegmentTree SegmentTree Int Int Val Val Val Val Bool
Node leftChild rightChild leftBound rightBound minimum maximum sume (modifier of sum) (modifier of equality)

findSegmentTreeSize :: Int -> Int

This function returns optimal size of segment tree - optimal is in case of power of two.

initialize :: Int -> Int -> SegmentTree

This function creates SegmentTree of specified size.

insertInitialNumbers :: Int -> Int -> [Val] -> SegmentTree -> SegmentTree

This function inserts some values into SegmentTree, values are given from beginning (left).

updateSetValueOnInterval :: Int -> Int -> Val -> SegmentTree -> SegmentTree

This function updates values (changes all values into given v) on specified inverval in SegmentTree

updateAddValueOnInterval :: Int -> Int -> Val -> SegmentTree -> SegmentTree

This function updates values (adds given v to all values) on specified inverval in SegmentTree

queryMinimumOnInterval :: Int -> Int -> SegmentTree -> Val

This function returns maximum value on specified inverval in SegmentTree

queryMaximumOnInterval :: Int -> Int -> SegmentTree -> Val

This function returns minimum value on specified inverval in SegmentTree

querySumOnInterval :: Int -> Int -> SegmentTree -> Val

This function returns sume of values on specified inverval in SegmentTree

getPartOfSegmentTree :: Int -> Int -> SegmentTree -> [Val]

With this function the values stored in SegmentTree or its structure can be printed in sensible way.

zero :: Val

This function returns zero of operations on SegmentTree

minusInf :: Val

This function returns minimum value (minus infinity) of operations on SegmentTree

plusInf :: Val

This function returns maximum value (plus infinity) of operations on SegmentTree

getLeftChild :: SegmentTree -> SegmentTree

This function returns LeftChild of specified Node of SegmentTree

getRightChild :: SegmentTree -> SegmentTree

This function returns RightChild of specified Node of SegmentTree

getLeftBound :: SegmentTree -> Int

This function returns LeftBound of specified Node of SegmentTree

getRightBound :: SegmentTree -> Int

This function returns RightBound of specified Node of SegmentTree

getMinimum :: SegmentTree -> Val

This function returns minimum on interval of specified Node of SegmentTree

getMaximum :: SegmentTree -> Val

This function returns maximum on interval of specified Node of SegmentTree

getSume :: SegmentTree -> Val

This function returns sume of values on interval of specified Node of SegmentTree

getModifierOfSum :: SegmentTree -> Val

This function returns modifier of sume of values on interval of specified Node of SegmentTree

getModifierOfEquality :: SegmentTree -> Bool

This function returns modifier of equality of values on interval of specified Node of SegmentTree